MNT #471 rename: avoid command line name confusion
