+++ profile_bluesky/startup/instrument/devices/__init__.py
@@ -34,10 +34,10 @@ from .trajectories import *
 from .usaxs_fly_scan import *
 
 # finally these area detectors
-from .alta import *
-from .blackfly import *
-from .dexela import *
-from .pilatus import *
+from .alta_module import *
+from .blackfly_module import *
+from .dexela_module import *
+from .pilatus_module import *
 # from .simdetector import *
 
 # and only when all devices are defined
